requirements.txt
Flask==3.1.0
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
